Refactor Q8 inference support modules
